refactor: Move segmentation and centroid settings to ObjectSettings
