+++ src/components/Tabs/Settings/ObjectSettings.tsx
@@ -1,12 +1,15 @@
+import { Checkbox } from "antd";
 import type { PresetsItem } from "antd/es/color-picker/interface";
 import React, { type ReactElement } from "react";
 import type { Color } from "three";
 
-import { DrawMode, KNOWN_CATEGORICAL_PALETTES, SelectionOutlineColorMode } from "src/colorizer";
+import { DrawMode, KNOWN_CATEGORICAL_PALETTES, SelectionOutlineColorMode, ViewMode } from "src/colorizer";
 import DropdownWithColorPicker from "src/components/Dropdowns/DropdownWithColorPicker";
 import type { SelectItem } from "src/components/Dropdowns/types";
+import LabeledSlider from "src/components/Inputs/LabeledSlider";
 import { SettingsContainer, SettingsItem } from "src/components/SettingsContainer";
 import ToggleCollapse from "src/components/ToggleCollapse";
+import { MAX_SETTINGS_SLIDER_WIDTH } from "src/constants";
 import { useViewerStateStore } from "src/state";
 
 import { DEFAULT_OUTLINE_COLOR_PRESETS, SETTINGS_GAP_PX } from "./constants";
@@ -16,6 +19,9 @@ const enum ObjectSettingsHtmlIds {
   EDGE_COLOR_SELECT = "edge-color-select",
   OUTLIER_OBJECT_COLOR_SELECT = "outlier-object-color-select",
   OUT_OF_RANGE_OBJECT_COLOR_SELECT = "out-of-range-object-color-select",
+  SHOW_CENTROIDS_SWITCH = "show-centroids-switch",
+  SHOW_SEGMENTATIONS_SWITCH = "show-segmentations-switch",
+  CENTROID_RADIUS_SLIDER = "centroid-radius-slider",
 }
 
 const DRAW_MODE_ITEMS = [
@@ -54,21 +60,31 @@ const OUTLINE_COLOR_MODE_ITEMS = [
 ] as const satisfies SelectItem[];
 
 export default function ObjectSettings(): ReactElement {
+  const centroidRadiusPx = useViewerStateStore((state) => state.centroidRadiusPx);
   const edgeColor = useViewerStateStore((state) => state.edgeColor);
   const edgeColorAlpha = useViewerStateStore((state) => state.edgeColorAlpha);
   const edgeMode = useViewerStateStore((state) => state.edgeMode);
   const outlierDrawSettings = useViewerStateStore((state) => state.outlierDrawSettings);
   const outlineColor = useViewerStateStore((state) => state.outlineColor);
   const outlineColorMode = useViewerStateStore((state) => state.outlineColorMode);
-  const outOfRangeDrawSettings = useViewerStateStore((state) => state.outOfRangeDrawSettings);
   const outlinePaletteKey = useViewerStateStore((state) => state.outlinePaletteKey);
+  const outOfRangeDrawSettings = useViewerStateStore((state) => state.outOfRangeDrawSettings);
+  const setCentroidRadiusPx = useViewerStateStore((state) => state.setCentroidRadiusPx);
   const setEdgeColor = useViewerStateStore((state) => state.setEdgeColor);
   const setEdgeMode = useViewerStateStore((state) => state.setEdgeMode);
   const setOutlierDrawSettings = useViewerStateStore((state) => state.setOutlierDrawSettings);
   const setOutlineColor = useViewerStateStore((state) => state.setOutlineColor);
   const setOutlineColorMode = useViewerStateStore((state) => state.setOutlineColorMode);
-  const setOutOfRangeDrawSettings = useViewerStateStore((state) => state.setOutOfRangeDrawSettings);
   const setOutlinePaletteKey = useViewerStateStore((state) => state.setOutlinePaletteKey);
+  const setOutOfRangeDrawSettings = useViewerStateStore((state) => state.setOutOfRangeDrawSettings);
+  const setShowCentroids = useViewerStateStore((state) => state.setShowCentroids);
+  const setShowSegmentations = useViewerStateStore((state) => state.setShowSegmentations);
+  const showCentroids = useViewerStateStore((state) => state.showCentroids);
+  const showSegmentations = useViewerStateStore((state) => state.showSegmentations);
+
+  // TODO: Remove when centroids are supported in 3D
+  const viewMode = useViewerStateStore((state) => state.viewMode);
+  const isDataset3d = viewMode === ViewMode.VIEW_3D;
 
   return (
     <ToggleCollapse label="Objects">
@@ -161,6 +177,46 @@ export default function ObjectSettings(): ReactElement {
             showColorPicker={outlierDrawSettings.mode === DrawMode.USE_COLOR}
           />
         </SettingsItem>
+        <SettingsItem label="Show segmentations" htmlFor={ObjectSettingsHtmlIds.SHOW_SEGMENTATIONS_SWITCH}>
+          <Checkbox
+            id={ObjectSettingsHtmlIds.SHOW_SEGMENTATIONS_SWITCH}
+            checked={showSegmentations}
+            onChange={(e) => setShowSegmentations(e.target.checked)}
+          />
+        </SettingsItem>
+        {/* TODO: Remove when centroids are supported in 3D */}
+        {!isDataset3d && (
+          <>
+            <SettingsItem label="Show centroids" htmlFor={ObjectSettingsHtmlIds.SHOW_CENTROIDS_SWITCH}>
+              <Checkbox
+                id={ObjectSettingsHtmlIds.SHOW_CENTROIDS_SWITCH}
+                checked={showCentroids}
+                onChange={(e) => setShowCentroids(e.target.checked)}
+              />
+            </SettingsItem>
+            <SettingsItem
+              label="Centroid radius"
+              htmlFor={ObjectSettingsHtmlIds.CENTROID_RADIUS_SLIDER}
+              labelStyle={{ marginTop: "1px" }}
+            >
+              <div style={{ maxWidth: MAX_SETTINGS_SLIDER_WIDTH }}>
+                <LabeledSlider
+                  id={ObjectSettingsHtmlIds.CENTROID_RADIUS_SLIDER}
+                  type="value"
+                  disabled={!showCentroids}
+                  step={0.5}
+                  value={centroidRadiusPx}
+                  onChange={setCentroidRadiusPx}
+                  minInputBound={0}
+                  maxInputBound={100}
+                  minSliderBound={1}
+                  maxSliderBound={10}
+                  numberFormatter={(value) => value?.toFixed(1)}
+                />
+              </div>
+            </SettingsItem>
+          </>
+        )}
       </SettingsContainer>
     </ToggleCollapse>
   );
